sniffer create and class for join data in orden.
segment start of header found/31/05/2016.

function size of packet and load header complete


capture of  packet cut in the network (03/06/2016)

organizacion trama of network is semi.full , problem in the 2 at 3 packet from three trama cut, trana not cut is full.(07/06/2016)

header ecg8 have a problem is out for a time, created a class for monitor and thread of process for each monitor in the networck.(08/06/2016).

grafichs  ecg1 2 3 good missing six grafich more.(09/06/2016)